Package test fixtures as pytest plugin 🫙
- Extract assertion fixtures and sentinels into filters.pytest_plugin so downstream libraries can consume them as proper injected fixtures
- Register via pytest11 entry point so pytest auto-discovers the plugin with no user configuration
- Slim conftest.py to internal-only helpers, confirming fixtures flow from the plugin
- Add test_pytest_plugin.py to verify the exported API and fixture injection
