+++ pyproject.toml
@@ -34,6 +34,9 @@ dependencies = [
     "regex>=2024",
 ]
 
+[project.entry-points."pytest11"]
+filters = "filters.pytest_plugin"
+
 [project.optional-dependencies]
 django = ["phx-filters-django"]
 iso = ["phx-filters-iso"]
